feat: A simple, light-weight food calorie analyzer app
The application accepts an image of the food that the user has consumed / is
going to consume, identifies that food, get a rough count of calories present in
it and inform the user how much time until they can eat their next meal.

The user can pass some notes on the contents of the food to help get a
"more" accurate estimation.